Edit profile data almost done and Refactor notifiers

+++ lib/src/ui/components/card_header_user_detail.dart
@@ -9,8 +9,8 @@ class CardHeaderUserDetail extends StatelessWidget {
   final String genderFormatted;
   final String creationDate;
   final bool isSelf;
-  final double? maxXp;
-  final double? xp;
+  final int? maxXp;
+  final int? xp;
   final Function? followUserCallback;
   final Function? editUserCallback;
